edicion y borrado de imagenes

+++ app/Http/Controllers/EmpleadoController.php
@@ -45,7 +45,8 @@ class EmpleadoController extends Controller
             $datosEmpleado['Foto'] = $request->file('Foto')->store('uploads','public');
         }
         Empleado::insert($datosEmpleado);
-        return response()->json($datosEmpleado);
+        // return response()->json($datosEmpleado);
+        return redirect('empleado')->with('mensaje','Empleado agregado con exito');
     }
 
     /**
@@ -107,8 +108,13 @@ class EmpleadoController extends Controller
     public function destroy($id)
     {
         //
-        Empleado::destroy($id);
-        return redirect('empleado');
+
+        $empleado = Empleado::findOrFail($id);
+
+        if(Storage::delete('public/'. $empleado->Foto)){
+            Empleado::destroy($id);
+        }
+        return redirect('empleado')->whith('mensaje','Empleado borrado');
 
 
     }
